Create list directive
